+++ example/open_route_service_example.dart
@@ -11,7 +11,8 @@ Future<void> main() async {
   const double endLng = -122.0792365;
 
   // Form Route between coordinates
-  final List<Coordinate> routeCoordinates = await client.getRouteCoordinates(
+  final List<Coordinate> routeCoordinates =
+      await client.directionsRouteCoordinatesGet(
     startCoordinate: Coordinate(latitude: startLat, longitude: startLng),
     endCoordinate: Coordinate(latitude: endLat, longitude: endLng),
   );